+++ components/bamboohr/src/actions/companyFiles/addCompanyFileCategory.ts
@@ -1,5 +1,6 @@
 import { action } from "@prismatic-io/spectral";
 import { createBambooClient } from "../../client";
+import { MESSAGES } from "../../constants";
 import { addCompanyFileCategoryExamplePayload } from "../../examplePayloads";
 import { addCompanyFileCategoryInputs } from "../../inputs";
 export const addCompanyFileCategory = action({
@@ -10,10 +11,14 @@ export const addCompanyFileCategory = action({
   inputs: addCompanyFileCategoryInputs,
   perform: async (context, params) => {
     const client = createBambooClient(params.connection, context.debug.enabled);
-    const { data } = await client.post("/v1/files/categories", [
+    const response = await client.post("/v1/files/categories", [
       params.categoryName,
     ]);
-    return { data };
+    const message =
+      response.status === 201
+        ? MESSAGES.CATEGORY_CREATED
+        : MESSAGES.CATEGORY_NOT_CREATED;
+    return { data: { message } };
   },
   examplePayload: addCompanyFileCategoryExamplePayload,
 });