+++ components/bamboohr/src/actions/employeeFiles/uploadEmployeeFile.ts
@@ -16,12 +16,13 @@ export const uploadEmployeeFile = action({
     formData.append("fileName", params.fileName);
     formData.append("share", params.share);
     formData.append("file", params.file.data, params.fileName);
-    await client.post(
+    const response = await client.post(
       `/v1/employees/${params.employeeId}/files`,
       formData.getBuffer(),
       { headers: formData.getHeaders() },
     );
-    return { data: null };
+    const location = response.headers.location || response.headers.Location;
+    return { data: { location } };
   },
   examplePayload: uploadEmployeeFileExamplePayload,
 });